First messy attempt at integrating pass with ember-fedora-adapter building off of previous fedora experiments.
