+++ app/controllers/login.js
@@ -8,7 +8,10 @@ export default Controller.extend({
     login() {
       this.set('loginFailed', false);
 
-      this.get('session').login(this.get('username')).then(() => {
+      let user = this.get('username');
+      let pass = this.get('password');
+
+      this.get('session').login(user, pass).then(() => {
         if (this.get('session.authenticated')) {
           let previousTransition = this.get('previousTransition');
 
@@ -21,6 +24,8 @@ export default Controller.extend({
         } else {
           this.set('loginFailed', true);
         }
+      }, () => {
+        this.set('loginFailed', true);
       });
     }
   }
