+++ context.jsonld
@@ -1,0 +1,63 @@
+{
+  "@context": {
+    "pass": "http://example.org/pass/",
+    "xsd": "http://www.w3.org/2001/XMLSchema#",
+
+    "Deposit": "pass:Deposit",
+    "Funder": "pass:Funder",
+    "Grant": "pass:Grant",
+    "Identifier": "pass:Identifier",
+    "Journal": "pass:Journal",
+    "Person": "pass:Person",
+    "Publisher": "pass:Publisher",
+    "Submission": "pass:Submission",
+    "User": "pass:User",
+    "Workflow": "pass:Workflow",
+    
+    "abstract": {"@id": "pass:abstract", "@type": "@id"},
+    "author": {"@id": "pass:author", "@type": "@id"},
+    "awardNumber": {"@id": "pass:awardNumber"},
+    "assignedId": {"@id": "pass:assignedId", "@type": "@id"},
+    "copis": {"@id": "pass:copis", "@container": "@set", "@type": "@id"},    
+    "creationDate": {"@id": "pass:creationDate", "@type": "xsd:dateTime"},    
+    "creator": {"@id": "pass:creator", "@type": "@id"},
+    "deposits": {"@id": "pass:deposits", "@container": "@set", "@type": "@id"},        
+    "doi": {"@id": "pass:doi", "@type": "@id"},    
+    "email": {"@id": "pass:email"},
+    "endDate": {"@id": "pass:endDate", "@type": "xsd:dateTime"}, 
+    "externalId": {"@id": "pass:externalId", "@type": "@id"},    
+    "funder": {"@id": "pass:funder", "@type": "@id"},
+    "grant": {"@id": "pass:grant", "@type": "@id"},
+    "grants": {"@id": "pass:grants", "@container": "@set", "@type": "@id"},    
+    "identifier": {"@id": "pass:identifier", "@type": "@id"},
+    "issue": {"@id": "pass:issue"},    
+    "ISSNs": {"@id": "pass:ISSNs", "@container": "@set", "@type": "@id"},    
+    "label": {"@id": "pass:label"},    
+    "journal": {"@id": "pass:journal", "@type": "@id"},
+    "journals": {"@id": "pass:journals", "@container": "@set", "@type": "@id"},
+    "name": {"@id": "pass:name"},
+    "nlmta": {"@id": "pass:nlmta"},
+    "oapCompliance": {"@id": "pass:oapCompliance"},
+    "orcid": {"@id": "pass:orcid"},    
+    "pi": {"@id": "pass:pi", "@type": "@id"},
+    "pmcParticipation": {"@id": "pass:pmcParticipation"},    
+    "projectName": {"@id": "pass:projectName"},
+    "publisher": {"@id": "pass:publisher"},    
+    "repo": {"@id": "pass:repo"},
+    "requested": {"@id": "pass:requested"},
+    "role": {"@id": "pass:role"},    
+    "status": {"@id": "pass:status"},
+    "startDate": {"@id": "pass:startDate", "@type": "xsd:dateTime"},
+    "step": {"@id": "pass:step"},
+    "steps": {"@id": "pass:steps"},    
+    "type": {"@id": "pass:type"},
+    "title": {"@id": "pass:title"},
+    "submissions": {"@id": "pass:submissions", "@container": "@set"},
+    "submittedDate": {"@id": "pass:submittedDate", "@type": "xsd:dateTime"},
+    "updatedDate": {"@id": "pass:updatedDate", "@type": "xsd:dateTime"},
+    "uri": {"@id": "pass:uri"},    
+    "username": {"@id": "pass:username"},
+    "volume": {"@id": "pass:volume"},
+    "workflows": {"@id": "pass:workflows", "@container": "@set", "@type": "@id"}    
+  }
+}
